feat(normalization): perform database normalization
- normalize the database upto a realistic database normalization
